+++ src/instance.c
@@ -322,6 +322,8 @@ lcb_error_t lcb_create(lcb_t *instance,
     lcb_set_view_timeout(obj, LCB_DEFAULT_VIEW_TIMEOUT);
     obj->rbufsize = LCB_DEFAULT_RBUFSIZE;
     obj->wbufsize = LCB_DEFAULT_WBUFSIZE;
+    obj->durability_timeout = LCB_DEFAULT_DURABILITY_TIMEOUT;
+    obj->durability_interval = LCB_DEFAULT_DURABILITY_INTERVAL;
 
     lcb_behavior_set_config_errors_threshold(obj, LCB_DEFAULT_CONFIG_ERRORS_THRESHOLD);
 
@@ -339,6 +341,7 @@ lcb_error_t lcb_create(lcb_t *instance,
     }
     obj->timers = hashset_create();
     obj->http_requests = hashset_create();
+    obj->durability_polls = hashset_create();
     /* No error has occurred yet. */
     obj->last_error = LCB_SUCCESS;
 
@@ -400,7 +403,21 @@ void lcb_destroy(lcb_t instance)
         }
         hashset_destroy(instance->timers);
     }
+
     lcb_connection_cleanup(&instance->connection);
+    if (instance->durability_polls) {
+        struct lcb_durability_set_st **dset_list;
+        lcb_size_t nitems = hashset_num_items(instance->durability_polls);
+        dset_list = (struct lcb_durability_set_st **)
+                hashset_get_items(instance->durability_polls, NULL);
+        if (dset_list) {
+            for (ii = 0; ii < nitems; ii++) {
+                lcb_durability_dset_destroy(dset_list[ii]);
+            }
+            free(dset_list);
+        }
+        hashset_destroy(instance->durability_polls);
+    }
 
     if (instance->vbucket_config != NULL) {
         vbucket_config_destroy(instance->vbucket_config);