+++ src/synchandler.c
@@ -202,6 +202,18 @@ static void observe_callback(lcb_t instance,
     lcb_maybe_breakout(instance);
 }
 
+static void durability_callback(lcb_t instance,
+                                const void *cookie,
+                                lcb_error_t error,
+                                const lcb_durability_resp_t *resp)
+{
+    struct user_cookie *c = (void *)instance->cookie;
+    restore_user_env(instance);
+    c->callbacks.durability(instance, cookie, error, resp);
+    restore_wrapping_env(instance, c, error);
+    lcb_maybe_breakout(instance);
+}
+
 static void unlock_callback(lcb_t instance,
                             const void *cookie,
                             lcb_error_t error,
@@ -243,6 +255,7 @@ static void restore_wrapping_env(lcb_t instance,
     instance->callbacks.http_data = http_data_callback;
     instance->callbacks.observe = observe_callback;
     instance->callbacks.unlock = unlock_callback;
+    instance->callbacks.durability = durability_callback;
 
     user->cookie = (void *)instance->cookie;
     user->retcode = error;