+++ tests/mock-unit-test.cc
@@ -62,6 +62,12 @@ void MockUnitTest::createConnection(lcb_t &instance)
     lcb_wait(instance);
 }
 
+void MockUnitTest::createConnection(HandleWrap& handle)
+{
+    lcb_t instance = NULL;
+    createConnection(handle, instance);
+}
+
 extern "C" {
     static void flags_store_callback(lcb_t,
                                      const void *,